added asynchronous annotation loading

+++ SystemyMobilne/Classes/Photos/SMPhotoScrollView.m
@@ -168,7 +168,7 @@
     return _zoomingView;
 }
 
-- (void)scrollViewDidEndZooming:(UIScrollView *)scrollView withView:(UIView *)view atScale:(float)scale
+- (void)scrollViewDidEndZooming:(UIScrollView *)scrollView withView:(UIView *)view atScale:(CGFloat)scale
 {
     
 }
